Moved session helpers out of session.
